+++ src/main/java/net/geforcemods/securitycraft/blockentities/BlockChangeDetectorBlockEntity.java
@@ -172,6 +172,7 @@ public class BlockChangeDetectorBlockEntity extends DisguisableBlockEntity imple
 		return disabled.get();
 	}
 
+	@Override
 	public boolean ignoresOwner() {
 		return ignoreOwner.get();
 	}