+++ src/main/java/net/geforcemods/securitycraft/blockentities/FloorTrapBlockEntity.java
@@ -124,6 +124,7 @@ public class FloorTrapBlockEntity extends DisguisableBlockEntity implements ITic
 		};
 	}
 
+	@Override
 	public boolean ignoresOwner() {
 		return ignoreOwner.get();
 	}