+++ src/main/java/net/geforcemods/securitycraft/blockentities/InventoryScannerBlockEntity.java
@@ -482,6 +482,7 @@ public class InventoryScannerBlockEntity extends DisguisableBlockEntity implemen
 		}
 	}
 
+	@Override
 	public boolean ignoresOwner() {
 		return ignoreOwner.get();
 	}