+++ src/main/java/net/geforcemods/securitycraft/blockentities/ProtectoBlockEntity.java
@@ -101,6 +101,7 @@ public class ProtectoBlockEntity extends DisguisableBlockEntity implements ITick
 		return disabled.get();
 	}
 
+	@Override
 	public boolean ignoresOwner() {
 		return ignoreOwner.get();
 	}